Remove old debugging code Move color coding
